Support start==end, linting, review comments
